+++ qmanager/modules/qmanager.cpp
@@ -304,36 +304,6 @@ out:
         flux_log_error (h, "%s: flux_respond_error", __FUNCTION__);
 }
 
-static void feasibility_request_cb (flux_t *h,
-                                    flux_msg_handler_t *w,
-                                    const flux_msg_t *msg,
-                                    void *arg)
-{
-    flux_future_t *f = nullptr;
-    const char *data = nullptr;
-
-    if (flux_request_decode (msg, nullptr, &data) < 0)
-        goto error;
-    if (!(f = flux_rpc (h, "sched-fluxion-resource.satisfiability", data, FLUX_NODEID_ANY, 0))) {
-        flux_log_error (h, "%s: flux_rpc (sched-fluxion-resource.satisfiability)", __FUNCTION__);
-        goto error;
-    }
-    if (flux_rpc_get (f, &data) < 0)
-        goto error;
-    if (flux_respond (h, msg, data) < 0) {
-        flux_log_error (h, "%s: flux_respond", __FUNCTION__);
-        goto error;
-    }
-    flux_log (h, LOG_DEBUG, "%s: feasibility succeeded", __FUNCTION__);
-    flux_future_destroy (f);
-    return;
-
-error:
-    if (flux_respond_error (h, msg, errno, flux_future_error_string (f)) < 0)
-        flux_log_error (h, "%s: flux_respond_error", __FUNCTION__);
-    flux_future_destroy (f);
-}
-
 static void params_request_cb (flux_t *h, flux_msg_handler_t *w, const flux_msg_t *msg, void *arg)
 {
     int saved_errno;
@@ -597,7 +567,6 @@ static void qmanager_destroy (std::shared_ptr<qmanager_ctx_t> &ctx)
 
 static const struct flux_msg_handler_spec htab[] = {
     {FLUX_MSGTYPE_REQUEST, "sched.resource-status", status_request_cb, FLUX_ROLE_USER},
-    {FLUX_MSGTYPE_REQUEST, "*.feasibility", feasibility_request_cb, FLUX_ROLE_USER},
     {FLUX_MSGTYPE_REQUEST, "*.params", params_request_cb, FLUX_ROLE_USER},
     FLUX_MSGHANDLER_TABLE_END,
 };
